requirements.txt
aiogram==2.22.2
aiohttp==3.8.6
bcrypt
PyJWT
babel
